+++ codegen/countries-to-continents.txt
@@ -1,255 +1,254 @@
-AS
-EU
-EU
-AS
-AS
-NA
-NA
-EU
-AS
-NA
-AF
-AN
-SA
-OC
-EU
-OC
-NA
-AS
-EU
-NA
-AS
-EU
-AF
-EU
-AS
-AF
-AF
-NA
-AS
-SA
-SA
-NA
-AS
-AN
-AF
-EU
-NA
-NA
-AS
-AF
-AF
-AF
-EU
-AF
-OC
-SA
-AF
-AS
-SA
-NA
-NA
-AF
-AS
-AS
-EU
-EU
-AF
-EU
-NA
-NA
-AF
-SA
-EU
-AF
-AF
-AF
-EU
-AF
-EU
-OC
-SA
-OC
-EU
-EU
-NA
-AF
-EU
-NA
-AS
-SA
-AF
-EU
-NA
-AF
-AF
-NA
-AF
-EU
-AN
-NA
-OC
-AF
-SA
-AS
-AN
-NA
-EU
-NA
-EU
-AS
-EU
-AS
-AS
-AS
-AS
-AS
-EU
-EU
-NA
-AS
-AS
-AF
-AS
-AS
-OC
-AF
-NA
-AS
-AS
-AS
-NA
-AS
-AS
-AS
-NA
-EU
-AS
-AF
-AF
-EU
-EU
-EU
-AF
-AF
-EU
-EU
-AF
-OC
-EU
-AF
-AS
-AS
-AS
-OC
-NA
-AF
-NA
-EU
-AF
-AS
-AF
-NA
-AS
-AF
-AF
-OC
-AF
-OC
-AF
-NA
-EU
-EU
-AS
-OC
-OC
-OC
-AS
-NA
-SA
-OC
-OC
-AS
-AS
-EU
-NA
-OC
-NA
-AS
-EU
-OC
-SA
-AS
-AF
-EU
-EU
-AF
-AS
-OC
-AF
-AF
-EU
-AS
-AF
-EU
-EU
-EU
-AF
-EU
-AF
-AF
-SA
-AF
-NA
-AS
-AF
-NA
-AF
-AN
-AF
-AS
-AS
-OC
-AS
-AF
-OC
-AS
-EU
-NA
-OC
-AS
-AF
-EU
-AF
-OC
-NA
-SA
-AS
-EU
-NA
-SA
-NA
-NA
-AS
-OC
-OC
-OC
-AS
-AF
-EU
-AF
-AF
-EU
-AF
--
--
--
-EU
-EU
-EU
-EU
-NA
-NA
-NA
-AF
+Asia
+Europe
+Europe
+Asia
+Asia
+NorthAmerica
+NorthAmerica
+Europe
+Asia
+NorthAmerica
+Africa
+Antarctica
+SouthAmerica
+Oceania
+Europe
+Oceania
+NorthAmerica
+Asia
+Europe
+NorthAmerica
+Asia
+Europe
+Africa
+Europe
+Asia
+Africa
+Africa
+NorthAmerica
+Asia
+SouthAmerica
+SouthAmerica
+NorthAmerica
+Asia
+Antarctica
+Africa
+Europe
+NorthAmerica
+NorthAmerica
+Asia
+Africa
+Africa
+Africa
+Europe
+Africa
+Oceania
+SouthAmerica
+Africa
+Asia
+SouthAmerica
+NorthAmerica
+NorthAmerica
+Africa
+Asia
+Asia
+Europe
+Europe
+Africa
+Europe
+NorthAmerica
+NorthAmerica
+Africa
+SouthAmerica
+Europe
+Africa
+Africa
+Africa
+Europe
+Africa
+Europe
+Oceania
+SouthAmerica
+Oceania
+Europe
+Europe
+NorthAmerica
+Africa
+Europe
+NorthAmerica
+Asia
+SouthAmerica
+Africa
+Europe
+NorthAmerica
+Africa
+Africa
+NorthAmerica
+Africa
+Europe
+Antarctica
+NorthAmerica
+Oceania
+Africa
+SouthAmerica
+Asia
+Antarctica
+NorthAmerica
+Europe
+NorthAmerica
+Europe
+Asia
+Europe
+Asia
+Asia
+Asia
+Asia
+Asia
+Europe
+Europe
+NorthAmerica
+Asia
+Asia
+Africa
+Asia
+Asia
+Oceania
+Africa
+NorthAmerica
+Asia
+Asia
+Asia
+NorthAmerica
+Asia
+Asia
+Asia
+NorthAmerica
+Europe
+Asia
+Africa
+Africa
+Europe
+Europe
+Europe
+Africa
+Africa
+Europe
+Europe
+Africa
+Oceania
+Europe
+Africa
+Asia
+Asia
+Asia
+Oceania
+NorthAmerica
+Africa
+NorthAmerica
+Europe
+Africa
+Asia
+Africa
+NorthAmerica
+Asia
+Africa
+Africa
+Oceania
+Africa
+Oceania
+Africa
+NorthAmerica
+Europe
+Europe
+Asia
+Oceania
+Oceania
+Oceania
+Asia
+NorthAmerica
+SouthAmerica
+Oceania
+Oceania
+Asia
+Asia
+Europe
+NorthAmerica
+Oceania
+NorthAmerica
+Asia
+Europe
+Oceania
+SouthAmerica
+Asia
+Africa
+Europe
+Europe
+Africa
+Asia
+Oceania
+Africa
+Africa
+Europe
+Asia
+Africa
+Europe
+Europe
+Europe
+Africa
+Europe
+Africa
+Africa
+SouthAmerica
+Africa
+NorthAmerica
+Asia
+Africa
+NorthAmerica
+Africa
+Antarctica
+Africa
+Asia
+Asia
+Oceania
+Asia
+Africa
+Oceania
+Asia
+Europe
+NorthAmerica
+Oceania
+Asia
+Africa
+Europe
+Africa
+Oceania
+NorthAmerica
+SouthAmerica
+Asia
+Europe
+NorthAmerica
+SouthAmerica
+NorthAmerica
+NorthAmerica
+Asia
+Oceania
+Oceania
+Oceania
+Asia
+Africa
+Europe
+Africa
+Africa
+Europe
+Africa
 
+
+
+Europe
+Europe
+Europe
+Europe
+NorthAmerica
+NorthAmerica
+NorthAmerica
+Africa
